requirements.txt
easyocr
pillow
numpy
torch
torchvision
transformers
timm
opencv-python
pdf2image
